requirements.txt
nltk
spacy
textblob
openai
python-dotenv
pinecone-client
tiktoken